Add Python environment setup logic
Automate the setup of a Python environment with a single `gmake init`
command.

Based on a provided Python version and a list of requirements, the
Python interpreter is installed (using pyenv), a virtual environment is
created and requirements are installed.

Additional `install-requirements` and `clear` rules allow the user to
recompile and install the requirements based on the `requirements.in`
and `requirements-dev.in` files, and to reset the Python environment,
respectively.